# docs/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Casa Natura API",
    "version": "1.0.0",
    "description": "Full API documentation for Casa Natura (public + admin endpoints).",
    "contact": {
      "name": "Casa Natura",
      "url": "https://www.deltagreen.it"
    }
  },
  "servers": [
    { "url": "https://www.deltagreen.it", "description": "Production" },
    { "url": "http://localhost:3000", "description": "Local" }
  ],
...
  "components": {
    "securitySchemes": {
      "cookieAuth": { "type": "apiKey", "in": "cookie", "name": "auth_token" }
    },
...
  },
  "paths": {
...
    "/api/checkout/create": {
      "post": {
        "summary": "Create Stripe payment",
        "tags": ["Checkout"],
        "requestBody": { "required": true, "content": { "application/json": { "schema": { "type": "object" } } } },
        "responses": { "200": { "description": "Stripe clientSecret + orderId" }, "409": { "description": "Out of stock" } }
      }
    },
...
  }
}